python_local/requirements.txt
lxml==4.6.3
Pillow==8.3.1
python-pptx==0.6.19
XlsxWriter==1.4.5